% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoveFun.R
\name{procesar_nc_a_excel}
\alias{procesar_nc_a_excel}
\title{Process NetCDF SST Data and Export to Excel}
\usage{
procesar_nc_a_excel(nc_path, excel_path)
}
\arguments{
\item{nc_path}{Character. Path to input NetCDF file.}

\item{excel_path}{Character. Path to output Excel file.}
}
\value{
Data frame with:
\itemize{
\item Date
\item Datetime
\item Mean SST
}
}
\description{
Reads a NetCDF file containing sea surface temperature data,
computes the spatial mean over time, and exports results to Excel.
}
\details{
Steps performed:
\enumerate{
\item Reads longitude, latitude, time, and temperature (thetao)
\item Converts time to datetime
\item Reshapes SST array into matrix
\item Computes mean SST per timestep
\item Saves output to Excel
}
}
\examples{
tab <- procesar_nc_a_excel("data.nc", "output.xlsx")

}
