% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoveFun.R
\name{shadow}
\alias{shadow}
\title{Shadow for ggplot}
\usage{
shadow(date, filling)
}
\arguments{
\item{date}{Date vector (YYYY-MM-DD)}

\item{filling}{Fill color for the rectangle}
}
\value{
A ggplot2 layer (geom_rect)
}
\description{
Shadow for ggplot
}
